+++ internal/common/common.go
@@ -10,7 +10,6 @@ type TestConfig struct {
 	ObservabilityBundleInstalled bool
 	SecurityBundleInstalled      bool
 	GatewayAPISupported          bool
-	IngressNginxSupported        bool
 	ARMNodePoolEnabled           bool
 }
 
@@ -25,7 +24,6 @@ func NewTestConfigWithDefaults() *TestConfig {
 		ObservabilityBundleInstalled: true,
 		SecurityBundleInstalled:      true,
 		GatewayAPISupported:          true,
-		IngressNginxSupported:        false,
 		ARMNodePoolEnabled:           false,
 	}
 }
@@ -37,7 +35,6 @@ func Run(cfg *TestConfig) {
 	runDNS(cfg.BastionSupported)
 	runMetrics(cfg)
 	runTeleport(cfg.TeleportSupported)
-	runHelloWorld(cfg.ExternalDnsSupported && cfg.IngressNginxSupported)
 	runHelloWorldGateway(cfg.GatewayAPISupported)
 	runScale(cfg.AutoScalingSupported)
 	runStorage()